Repeated Tackle Stun Resistance (#10845)
This PR is follow up to #9412 doing the following:
- Tackling someone as a xeno cannot last longer than 20s
- Whenever the tackle duration was reduced because the new application
would have exceeded the allowed 20s, the tackle sfx pitch is higher
- To accommodate above, tackle sfx pitch now does not vary (other than
for the higher pitch above)
- Stun and knockdown end for the hauled mob at the start of a haul

<details>
<summary>Initial implementation of this was doing the following:
</summary>

- Tackling someone who is already stunned now builds a resistance to
future tackle stun applications
- Currently the rate of accumulation is 7% per tackle but likely will
change
- Ever the stun ends the resistance is reset
- https://youtu.be/1MdVms341kA
</details>

The intention of this PR is to discourage tackle dragging as a nearly no
risk alternative to hauling. The goal is to allow marine agency when
getting transported. For hauling this is struggling with a weapon. Now
for tackle dragging it cannot be applied indefinitely without the marine
regaining control every 20s.

The end of stun/knockdown when hauled is to mitigate the possibility
that maybe someone would repeatedly haul and unhaul someone w/o that
person having a chance to struggle immediately when hauled. I don't
think there should be any other effects to this change (A hug is a
paralyze for example).

20s: https://youtu.be/6HcECoFK2xs

:cl: Drathek
balance: Xenomorph tackles now cannot stun longer than 20s.
balance: Hauling now ends any stun/knockdown on the hauled mob (Not
hugger paralyze)
add: Xenomorph tackle sfx has a higher pitch if the new duration was
altered because of the 20s limit (otherwise does not vary now)
/:cl: